requirements.txt
# Django and DRF
django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-environ==0.11.2

# GraphQL
graphene-django==3.2.0
django-graphql-jwt==0.4.0
django-filter==23.5

# Database
psycopg2-binary==2.9.9
redis==5.0.1

# Celery
celery==5.3.6
flower==2.0.1

# API and OpenAI
openai==1.12.0
requests==2.31.0
python-dotenv==1.0.0
python-magic==0.4.27

# Image Processing
Pillow==10.2.0

# Mobile Client
kivy==2.3.1
kivymd==1.2.0
sdl2_ttf==2.0.15

# Development Tools
black==23.11.0
flake8==6.1.0
pytest==7.4.3
pytest-django==4.7.0

# Documentation
drf-yasg==1.21.7

# Production
gunicorn==21.2.0
whitenoise==6.6.0 